backend/requirements.txt
# ==================== WEB FRAMEWORK ====================
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
python-multipart>=0.0.6

# ==================== VALIDATION & SETTINGS ====================
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# ==================== DATABASE ====================
aiosqlite>=0.19.0
alembic>=1.13.0
sqlalchemy[asyncio]>=2.0.25

# ==================== TELEGRAM ====================
aiogram>=3.4.0,<4.0.0

# ==================== HTTP CLIENT ====================
httpx>=0.26.0,<1.0.0

# ==================== RATE LIMITING ====================
slowapi>=0.1.9

# ==================== SECURITY ====================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# ==================== RETRY ====================
tenacity>=8.2.0

# ==================== LOGGING ====================
loguru>=0.7.0

# ==================== UTILS ====================
python-dotenv>=1.0.0
email-validator>=2.1.0

# ==================== EXPORT (PDF/EXCEL) ====================
reportlab>=4.0.0
openpyxl>=3.1.0
